Changes which make working with exceptions a bit more pleasant (and visisble at that) 	* html/includes/exceptions/textexceptionrendering.class.php: deprecated var declarators 	* html/includes/xarCore.php: madness again, E_STRICT and exceptions dont like eachother 	* html/includes/xarException.php: register a poor mans exception handler 	* html/includes/xarMod.php: reference errors 	* html/includes/xarTableDDL.php: check wether xar_tables table exist to prevent unnecessary errors

+++ html/includes/xarTableDDL.php
@@ -330,15 +330,15 @@ function xarDBDropTable($tableName, $databaseType = NULL)
     $metaTable = $systemPrefix . '_tables';
     if ($tableName != $metaTable) {
         $dbconn =& xarDBGetConn();
-        //$dbInfo = $dbconn->getDatabaseInfo();
-        //if($dbInfo->hasTable($metaTable)) {
+        $dbInfo = $dbconn->getDatabaseInfo();
+        if($dbInfo->hasTable($metaTable)) {
             $query = "DELETE FROM $metaTable WHERE xar_table=?";
             // This doesnt have to be fatal
             //try {
                 $result =& $dbconn->Execute($query,array($tableName));
             //} catch(Exception $e) {
             //}
-            //}
+        }
     }
 
     switch($databaseType) {
